refactor along feature lines
