+++ examples/counter/src/app/app.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgReduxModule, DevToolsExtension } from 'ng2-redux';
-import { CounterActions } from '../actions/counter.actions';
-import { SearchActions } from '../actions/search.actions';
+import { CounterActions } from '../counter/counter.actions';
+import { SearchActions } from '../search/search.actions';
 import { AppComponent } from './app.component';
-import { CounterInfoComponent } from '../components/counter-info.component';
-import { CounterComponent } from '../components/counter.component';
-import { SearchComponent } from '../components/search.component';
-import { RandomNumberService } from '../services/random-number.service';
+import { CounterInfoComponent } from '../counter/counter-info.component';
+import { CounterComponent } from '../counter/counter.component';
+import { SearchComponent } from '../search/search.component';
+import { RandomNumberService } from '../common/random-number.service';
 
 @NgModule({
   imports: [
