+++ examples/counter/src/common/random-number.service.ts
