+++ examples/counter/src/counter/counter-info.component.html
@@ -1,0 +1,6 @@
+<ul>
+  <li>{{ funcCounter$ | async }}</li>
+  <li>{{ stringKey$ | async }}</li>
+  <li>{{ counterX2$ | async }}</li>
+  <li>{{ foo?.x }} - {{ foo?.y }}</li>
+<ul>
