+++ examples/counter/src/counter/counter.actions.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { NgRedux } from 'ng2-redux';
 import { IAppState } from '../store';
-import { RandomNumberService } from '../services/random-number.service';
+import { RandomNumberService } from '../common/random-number.service';
 
 /**
  * Action creators in Angular 2. We may as well adopt a more
