+++ examples/counter/src/counter/counter.component.html
@@ -1,0 +1,10 @@
+<p>
+  Clicked: {{ counter$ | async }} times
+  <button (click)="actions.increment()">+</button>
+  <button (click)="actions.decrement()">-</button>
+  <button (click)="actions.incrementIfOdd()">Increment if odd</button>
+  <button (click)="actions.incrementAsync(2222)">Increment async</button>
+  <button (click)="actions.randomize()">Set to random number</button>
+</p>
+<p>foo$: {{ foo$ | async | json }}</p>
+<p>bar$: {{ bar$ | async }}</p>
