+++ examples/counter/src/search/search.component.html
@@ -1,0 +1,9 @@
+<input
+  id='search-input'
+  type="text"
+  class="search"
+  #field (keyup)="handleKeyUp(field.value)"/>
+<p>Number of characters (from subscription): {{ numChars }}</p>
+<p>Number of characters (from async pipe): {{ numChars$ | async }}</p>
+<p>You entered: {{ search$ | async }}<p>
+
